Move Responses section components to responses app router folder
