+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/responses/[statusFilter]/components/ConfirmReceiptStatus.tsx
@@ -1,7 +1,7 @@
 "use client";
 import React from "react";
 import { useTranslation } from "@i18n/client";
-import { ExclamationText } from "../shared";
+import { ExclamationText } from "../../../../../../../../../components/clientComponents/form-builder/app/shared";
 import { VaultStatus } from "@lib/types";
 import { getDaysPassed } from "@lib/client/clientHelpers";
