+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/responses/[statusFilter]/components/Dialogs/DownloadDialog.tsx
@@ -1,6 +1,9 @@
 "use client";
 import React, { useEffect } from "react";
-import { Dialog, useDialogRef } from "../../shared";
+import {
+  Dialog,
+  useDialogRef,
+} from "../../../../../../../../../../components/clientComponents/form-builder/app/shared";
 import { useTranslation } from "@i18n/client";
 import { Button } from "@clientComponents/globals";
 import { logMessage } from "@lib/logger";
