+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/responses/[statusFilter]/components/DownloadResponseStatus.tsx
@@ -2,7 +2,7 @@
 import { formatDate } from "@lib/client/clientHelpers";
 import React from "react";
 import { useTranslation } from "@i18n/client";
-import { ExclamationText } from "../shared";
+import { ExclamationText } from "../../../../../../../../../components/clientComponents/form-builder/app/shared";
 import { getDaysPassed } from "@lib/client/clientHelpers";
 import { VaultStatus } from "@lib/types";
